Split out single device polling to a Laravel Job Probably totally broken :)
